changed: RND is now 100% compatible with QB in -lang qb mode (patch written by counting_pine)
r4220

+++ FreeBASIC/src/rtlib/fb_math.h
@@ -32,7 +32,7 @@
 #ifndef __FB_MATH_H__
 #define __FB_MATH_H__
 
-FBCALL double       fb_Rnd              ( int n );
+FBCALL double       fb_Rnd              ( float n );
 FBCALL void         fb_Randomize        ( double seed, int algorithm );
 FBCALL int          fb_SGNSingle        ( float x );
 FBCALL int          fb_SGNDouble        ( double x );
